// src/types/DetailState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DetailState = { sharpenAmount: number, sharpenRadius: number, sharpenDetail: number, sharpenMasking: number, nrLuminance: number, nrLumaDetail: number, nrLumaContrast: number, nrColor: number, nrColorDetail: number, hotPixelRemoval: boolean, };